% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_polecat}
\alias{download_polecat}
\title{Download POLECAT}
\usage{
download_polecat(local_dir, years = NULL, verbose = TRUE)
}
\arguments{
\item{local_dir}{Which local directory to download the data to?}

\item{years}{Do you want to download data for only specific years? The
default (NULL) will download all available data. Otherwise, specify a
numeric vector, e.g. \code{2023}, \code{2022:2023}, or \code{c(2021, 2023)}.}

\item{verbose}{Print progress messages as files are being downloaded?}
}
\value{
Data will be downloaded to \code{local_dir}.
}
\description{
Download the POLECAT data from dataverse.
}
