% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_dyadic}
\alias{filter_dyadic}
\title{Filter to dyadic events}
\usage{
filter_dyadic(x)
}
\arguments{
\item{x}{A data frame of POLECAT events.}
}
\value{
A filtered data frame.
}
\description{
Keep only events where both source and target actors are identified.
}
\examples{
data(polecat_sample)
dyadic <- filter_dyadic(polecat_sample)
}
