% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_interstate}
\alias{filter_interstate}
\title{Filter to interstate events}
\usage{
filter_interstate(x)
}
\arguments{
\item{x}{A data frame of POLECAT events.}
}
\value{
A filtered data frame.
}
\description{
Keep only events where the source and target are in different countries.
}
\examples{
data(polecat_sample)
foreign <- filter_interstate(polecat_sample)
}
