% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_intrastate}
\alias{filter_intrastate}
\title{Filter to intrastate events}
\usage{
filter_intrastate(x)
}
\arguments{
\item{x}{A data frame of POLECAT events.}
}
\value{
A filtered data frame.
}
\description{
Keep only events where the source and target are in the same country.
}
\examples{
data(polecat_sample)
domestic <- filter_intrastate(polecat_sample)
}
