% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state-dataverse.R
\name{get_dataverse_state}
\alias{get_dataverse_state}
\title{Get Dataverse POLECAT file state}
\usage{
get_dataverse_state()
}
\value{
A data frame with columns: \code{label}, \code{id},
  \code{version}, \code{creationDate}, and \code{data_year}.
}
\description{
Query the Harvard Dataverse API and return a data frame describing the
POLECAT data files currently available for download.
}
\examples{
\dontrun{
  get_dataverse_state()
}
}
