% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{get_polecat_opts}
\alias{get_polecat_opts}
\title{Get rpolecat options}
\usage{
get_polecat_opts()
}
\value{
A named list of option values.
}
\description{
Return the current values of all rpolecat options.
}
\examples{
get_polecat_opts()
}
