% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list-files.R
\name{list_dataverse_files}
\alias{list_dataverse_files}
\title{List Dataverse POLECAT files}
\usage{
list_dataverse_files()
}
\value{
Invisibly returns a data frame of file information (same as
  \code{\link{get_dataverse_state}}).
}
\description{
Query the Harvard Dataverse and print a formatted summary of the POLECAT
files available for download.
}
\examples{
\dontrun{
  list_dataverse_files()
}
}
