% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{modes}
\alias{modes}
\title{Event type modes}
\format{
An object of class \code{data.frame} with 73 rows and 2 columns.
}
\usage{
modes
}
\description{
Event type modes
}
\keyword{datasets}
