% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{quad_categories}
\alias{quad_categories}
\title{Quad categories}
\format{
A data frame with 16 rows and 3 columns:
\describe{
  \item{event_type}{PLOVER event type}
  \item{quad_category}{One of: Verbal Cooperation, Material Cooperation,
    Verbal Conflict, Material Conflict}
  \item{goldstein}{Approximate Goldstein conflict-cooperation score}
}
}
\description{
A data frame mapping PLOVER event types to quad categories (Verbal
Cooperation, Material Cooperation, Verbal Conflict, Material Conflict)
with associated Goldstein scale scores.
}
