% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpolecat-package.R
\docType{package}
\name{rpolecat-package}
\alias{rpolecat-package}
\alias{_PACKAGE}
\alias{rpolecat}
\title{Download and Transform the POLECAT Event Data}
\description{
Download the POLECAT data from Dataverse.
}
\author{
\strong{Maintainer}: Andreas Beger \email{andy@basilanalytics.com} (\href{https://orcid.org/0000-0003-1883-3169}{ORCID})

Authors:
\itemize{
  \item Benjamin E. Bagozzi \email{bagozzib@udel.edu}
}

}
