% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{set_polecat_opts}
\alias{set_polecat_opts}
\title{Set rpolecat options}
\usage{
set_polecat_opts(data_dir = NULL, use_db = NULL, db_path = NULL)
}
\arguments{
\item{data_dir}{Path to local data directory.}

\item{use_db}{Logical; use a database backend?}

\item{db_path}{Path to the database file.}
}
\value{
Invisibly returns the new option values.
}
\description{
Programmatically set individual rpolecat options.
}
\examples{
\dontrun{
  set_polecat_opts(data_dir = "~/data/polecat")
}
}
