% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{unset_polecat_opts}
\alias{unset_polecat_opts}
\title{Unset rpolecat options}
\usage{
unset_polecat_opts()
}
\value{
Invisibly returns \code{TRUE}.
}
\description{
Remove all rpolecat options from the current session.
}
\examples{
unset_polecat_opts()
}
